Reinitialization readme formatting
