+++ README.md
@@ -1,6 +1,8 @@
 # react-quick-reactions
 
-A small popup component for quick reactions and emojis, à la GitHub's reaction popup or Facebook's "likes".
+Under construction! Currently a work-in-progress.
+
+A lightweight popup component for quick reactions and emojis, à la GitHub's reaction popup or Facebook's "likes".
 
 # Installation
