+++ package.json
@@ -1,52 +1,48 @@
 {
   "name": "react-quick-reactions",
-  "version": "0.1.0",
-  "private": false,
+  "version": "0.2.0",
   "description": "A popup emoji reaction component for React.",
-  "main": "src/index.tsx",
-  "license": "MIT",
+  "private": false,
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
+  "files": [
+    "dist"
+  ],
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "rollup": "rollup -c"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DylanBlair/react-quick-reactions.git"
+  },
   "keywords": [
     "react",
     "reactions",
     "react-component",
     "quick-reactions",
-    "emoji-reactions"
+    "emoji-reactions",
+    "front-end",
+    "UI"
   ],
-  "homepage": "https://github.com/Dylandbl/react-quick-reactions#readme",
+  "author": "Dylan Blair <dylandbl.dev@gmail.com>",
+  "license": "MIT",
   "bugs": {
-    "url": "https://github.com/Dylandbl/react-quick-reactions/issues"
+    "url": "https://github.com/DylanBlair/react-quick-reactions/issues"
   },
-  "repository": "github:Dylandbl/react-quick-reactions",
-  "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "homepage": "https://github.com/DylanBlair/react-quick-reactions#readme",
+  "publishConfig": {
+    "registry": "https://www.npmjs.com/package/react-quick-reactions"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-typescript": "^8.3.0",
     "@types/react": "^17.0.39",
+    "react": "^17.0.2",
+    "rollup": "^2.67.3",
+    "rollup-plugin-dts": "^4.1.0",
     "typescript": "^4.5.5"
   }
 }
