--- src/index.tsx
@@ -1,2 +1,0 @@
-export { default as ReactionPopover } from "./components/ReactionPopover";
-export { default as PopoverProps } from "./types/PopoverProps";
